requirements.txt
astroid==2.2.5
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-filter==2.2.0
django-phonenumber-field==3.0.1
djangorestframework==3.10.2
idna==2.8
instamojo-wrapper==1.2.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.2
phonenumbers==8.10.14
pylint==2.3.1
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
